Change from basic js to jquery
